Implemented basic mesh simplification algorithm
This version of the algorithm always simplifies meshes so that they contain a
subset of the vertices contained in the original mesh--it never creates new
vertices. It also ignores texcoords and normals. And the interface with
pycollada is still a little clunky (it just simplifies the first primitive from
the first geometry and overwrites that primitive with the simplified one, then
saves the entire result in a new file). Arbitrarily, it always simplifies away
4/5 of the vertices right now.

I've verified that this code works with the duck model, at least.