Add boundary constraints
Modified the error metric so that vertices on boundaries (edges that are
adjacent to only one face) stay close to those boundaries. This helps a lot when
meshes are composed of multiple disconnected components that each have open
boundaries. In the sample DAE files I've been testing on, this has been a pretty
common case.